Modify the generateSquare() location to use all the area above minRadius
Original implementation added the minimum value onto all randomly generated
values, this mean that the min area was never used even when the second axis
would take the value outside the radius. This resulted in a pattern of invalid
location forming a + shape over the entire map like so:

VVVVIIVVVV
VVVVIIVVVV
IIIIIIIIII
IIIIIIIIII
VVVVIIVVVV
VVVVIIVVVV

This change modifieds the generation to use completely random values where
only values generated where both the X and the Z coordinates are considered
invalid and forces another randomly generated value to be chosed on one of
the 2 axis.